Refactor: react.fc 제거 및 error,loading 처리 명확하게 구분
