+++ src/feature/channel/api/usePlaylistInfo.ts
@@ -1,13 +1,6 @@
 import axiosInstance from '@/shared/lib/axios/axiosInstance'
 import { useQuery } from '@tanstack/react-query'
-
-export interface IPlayList {
-  id: number
-  created_at: string
-  user_id: number
-  name: string
-  access: string
-}
+import { IPlayList } from '../tpye/IChannel'
 
 export const fetchPlayList = async (userId: number): Promise<IPlayList[]> => {
   const response = await axiosInstance.get(`playlist?user_id=eq.${userId}&order=created_at.desc`)
