+++ src/feature/channel/tpye/IChannel.ts
@@ -7,3 +7,11 @@ export interface IPlayListIemProps {
 export interface ITabMenuProps {
   onChangeTab: (tab: 'video' | 'playlist') => void
 }
+
+export interface IPlayList {
+  id: number
+  created_at: string
+  user_id: number
+  name: string
+  access: string
+}
