+++ src/feature/video/api/useUserImg.ts
@@ -1,6 +1,15 @@
 import { useQuery } from '@tanstack/react-query'
 import { supabase } from '@/shared/lib/supabase/supabaseClient'
 
+export const useUserImgInfo = (userId: number) => {
+  return useQuery({
+    queryKey: ['userImg', userId],
+    queryFn: () => getUserInfo(userId),
+    enabled: !!userId,
+    staleTime: 1000 * 60 * 5,
+  })
+}
+
 // ✅ 유저의 이미지를 가져옴
 const getUserInfo = async (id: number) => {
   const { data, error } = await supabase.from('users').select('img, name').eq('id', id).single()
@@ -11,13 +20,4 @@ const getUserInfo = async (id: number) => {
   }
 
   return [data.img, data.name]
-}
-
-export const useUserInfo = (userId: number) => {
-  return useQuery({
-    queryKey: ['userImg', userId],
-    queryFn: () => getUserInfo(userId),
-    enabled: !!userId,
-    staleTime: 1000 * 60 * 5,
-  })
 }
